Moves relate.interp -> relate
